Сделано обновление БД от старой версии к новой

+++ app/src/main/java/com/gawk/voicenotes/models/Note.java
@@ -22,7 +22,7 @@ public class Note implements Serializable {
         this.date = null;
     }
 
-    public Note(int _id, String _text_note, Date _date) {
+    public Note(long _id, String _text_note, Date _date) {
         this.id = _id;
         this.text_note = _text_note;
         this.date = _date;
